Added support for Query Guards. Ex: `/?<user..>`
This allows multiple query parameters to be added to the list of
parameters in the documentation.
It also allows custom Query Guards to be created.
Code for this is included in documentation.
It also functions as a test.

There might be a problem with this code if more structs use the
same name in different modules and both implement `JsonSchema`.
The reference system might not work in that case.