Upgrade to swagger-ui 3.34.0
In the previous swagger-ui version, recursive Schemas (Schemas referring
to themselves) were leading to infinite loops when rendering the
documentation. This seems to be fixed in the more recent swagger-ui
version.

The new files are the result of downloading the current swagger-ui
release from swagger-api/swagger-ui and then running `npm install`
followed by `npm run build` and extracting only the files that were
already found here from the `dist/` directory.