fix deserialisation for indexmap case
The 'extensions' fields of various structs
within okapi/src/openapi3.rs was not
deserialized correctly.

This commit fixes the behavior if the Object is
an indexmap. Using rocket_okapi enables the feature
'preserve_order' by default and hence an indexmap is used.

If the feature 'preserve_order' is disables, this commit should
not change the behavior.

Belongs to https://github.com/GREsau/okapi/issues/67

Signed-off-by: Philippe Schaaf <philippe.schaaf@secunet.com>